+++ src/wishes/wishes.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { Wish } from './entities/wish.entity';
+import { CreateUserDto } from '../users/dto/create-user.dto';
 
 @Injectable()
 export class WishesService {
@@ -25,11 +26,27 @@ export class WishesService {
     return this.wishRepository.findOneBy({ id });
   }
 
-  update(id: string, updateWishDto: UpdateWishDto) {
+  update({
+    executor,
+    id,
+    updateWishDto,
+  }: {
+    executor: CreateUserDto;
+    id: string;
+    updateWishDto: UpdateWishDto;
+  }) {
+    if (String(executor.id) !== String(id)) {
+      throw new ForbiddenException();
+    }
+
     return this.wishRepository.update({ id }, updateWishDto);
   }
 
-  remove(id: string) {
+  remove({ executor, id }: { executor: CreateUserDto; id: string }) {
+    if (String(executor.id) !== String(id)) {
+      throw new ForbiddenException();
+    }
+
     return this.wishRepository.delete({ id });
   }
 }
